% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date.R
\name{get_date}
\alias{get_date}
\title{Get Today's Date!}
\usage{
get_date(x, y)
}
\arguments{
\item{x}{x must be a char}

\item{y}{y must be a num}
}
\value{
returns today's date
}
\description{
Get Today's Date!
}
\examples{
get_date()
}
