% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model function}
\usage{
my_lm(my_fml, my_data)
}
\arguments{
\item{my_fml}{'formula' class object.}

\item{my_data}{Input data frame.}
}
\value{
A table of coeffiencts for the linear regression, which contains `Estimate`, `Std. Error`, `t value`, and `Pr(>|t|)`.
}
\description{
This function fits a linear model.
}
\examples{
data(my_gapminder)
my_lm(my_fml = pop ~ gdpPercap, my_data = my_gapminder)

}
\keyword{inference}
