% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation Function}
\usage{
my_rf_cv(k)
}
\description{
This function performs a Random Forest Cross-Validation in R
}
\details{
@param k number of folds

 @return a numeric with the cross-validation error;

 @examples
 my_rf_cv(k = 5)

 @export
}
