% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{t-test function}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data to be tested.}

\item{alternative}{Charater string specifying the alternative hypothesis. It is one of "two.sided", "less", or "greater".}

\item{mu}{Numeric input indicating the null hypothesis value of the mean.}
}
\value{
A List of test_stat (the numeric test statistic), df (degrees of freedom),  p_value (the numeric p-value),  and alternative (the value of the parameter alternative).
}
\description{
This function performs a one sample t-test.
}
\examples{
my_t_test(rnorm(100, mean = 0, sd = 1), mu = 0, alternative = "less")
my_t_test(c(1, 2, 3, 4, 5), mu = 3, alternative = "greater")

}
\keyword{Inferential}
\keyword{Statistic}
