% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-Nearest Neighbors Cross-Validation Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\description{
This function performs a k-Nearest Neighbors Cross-Validation in R
}
\details{
@param train Input data frame to be validated
 @param cl true class value of the training data
 @param k_nn integer representing the number of neighbors
 @param k_cv integer representing the number of folds

 @return a list with objects, which has class: a vector of the predicted class Y_i for all observations,
 and cv_err: a numeric with the cross-validation misclassification error;

 @examples
 my_knn_cv(train = my_iris[, -5], cl = my_iris$Species, k_nn = 5, k_cv = 5)

 @import class

 @export
}
